Implemented the Conway's Game Of Life simulation

+++ source/ProjectAutomata/ProjectAutomata/IO/FontManager.cpp
@@ -58,5 +58,5 @@ Font* FontManager::getFont(char* name)
 
 int FontManager::getFontCount()
 {
-	return fonts.size();
+	return (int)fonts.size();
 }
